+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * <kbd>C-c M-f</kbd> Select a function from the current namespace using IDO and insert into the REPL buffer.
+* `cider-read-and-eval` now supports completion and keeps history.
 
 ## 0.4.0 / 2013-12-03